incorporate some minor fixes
git-svn-id: https://svn.r-project.org/R/trunk@57571 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/extra/tre/regcomp.c
@@ -39,7 +39,7 @@ tre_regncomp(regex_t *preg, const char *regex, size_t n, int cflags)
   if (TRE_MB_CUR_MAX == 1)
 #endif /* TRE_MULTIBYTE */
     {
-      unsigned int i;
+      size_t i;
       const unsigned char *str = (const unsigned char *)regex;
       tre_char_t *wstr = wregex;
 
@@ -50,7 +50,7 @@ tre_regncomp(regex_t *preg, const char *regex, size_t n, int cflags)
 #if TRE_MULTIBYTE
   else
     {
-      int consumed;
+      size_t consumed;
       tre_char_t *wcptr = wregex;
 #ifdef HAVE_MBSTATE_T
       mbstate_t state;
@@ -71,11 +71,11 @@ tre_regncomp(regex_t *preg, const char *regex, size_t n, int cflags)
 		  return REG_BADPAT;
 		}
 	      break;
-	    case -1:
+	    case (size_t)-1:
 	      DPRINT(("mbrtowc: error %d: %s.\n", errno, strerror(errno)));
 	      xfree(wregex);
 	      return REG_BADPAT;
-	    case -2:
+	    case (size_t)-2:
 	      /* The last character wasn't complete.  Let's not call it a
 		 fatal error. */
 	      consumed = n;
