+++ src/extra/tre/tre-match-approx.c
@@ -523,7 +523,7 @@ tre_tnfa_run_approx(const tre_tnfa_t *tnfa, const void *string, int len,
 
 	    /* Pop the first item off the deque. */
 	    reach_p = *deque_start;
-	    id = reach_p - reach_next;
+	    id = (int)(reach_p - reach_next);
 	    depth = reach_p->depth;
 
 	    /* Compute cost at current depth. */
